First cut of parser script
This seems to work well for old reports (such as from 2019-09) but not newer reports (such as 2022-11) that have a different format.  The scirpt is using the pdfreader Python package and I'm not sure why pdfreader can't read new reports but I'll try to find an alternative way to read the reports.