Tweaking processing
I found a glitch in processing after I switched to pdftotext but it was easy to fix.  The problem was that the `reference interval` could contain multiple imbedded blanks that look weird, throw off the tabular columns, etc.

Here's an example from the old report I was using for testing:

  Date        Test                   Result  Flag  Units        Reference Interval  Lab
  2019-09-24  BUN                    14            mg/dL        6 - 24              01

This looks fine but the next report is has the weird blanks:

  Date        Test                   Result  Flag  Units        Reference Interval                             Lab
  2020-01-23  BUN                    10            mg/dL        6                                   - 24       01

Fudge that!  It was easy to slip in a little code to replace multiple spaces with a single space for beautiful output:

  Date        Test                   Result  Flag  Units        Reference Interval  Lab
  2020-01-23  BUN                    10            mg/dL        6 - 24              01