Tweaking reference interval
I have tweaked reference intervals in the past but wanted to improve it especially in the case of 'min - max'.  I was doing the following:
  - substitute multiple blanks with a single blank
  - replacing '-' with ' - '

I think some of the reports originally had 'min-max' and some had 'min - max' and because of the order I was doing the tweaks, I could end up with 'min  -  max' which was annoying.  I decided to swap the order of the tweaks so I would get 'min - max' more reliably and with only one space between non-spaces.